modifier final for class added. Remove label fom TapBaController. Change insets between photos. Extentions for UIColor moved in new directory. modifier privat for all ib outlets and actions added
