+++ ImageFeed/Servises/FavoritesServise.swift
@@ -1,0 +1,12 @@
+//
+//  FavoritesServise.swift
+//  ImageFeed
+//
+//  Created by Maksim Zakharov on 07.09.2024.
+//
+
+import Foundation
+
+class FavoritesServise {
+    
+}
