+++ ImageFeed/ViewConrollers/ImagesListView/ImagesListViewController.swift
@@ -86,8 +86,6 @@ extension ImagesListViewController: UITableViewDelegate {
         let cellHeight = image.size.height * scale + imageInsets.top + imageInsets.bottom
         return cellHeight
     }
-    
-    
 }
 
 extension ImagesListViewController {
