+++ ImageFeed/ViewConrollers/profileView/ProfileViewController.swift
@@ -8,7 +8,7 @@
 import Foundation
 import UIKit
 
-class ProfileViewController: UIViewController {
+final class ProfileViewController: UIViewController {
     // MARK: - IB Outlets
     
     // MARK: - Public Properties
@@ -161,28 +161,3 @@ class ProfileViewController: UIViewController {
     }
 }
 
-// MARK: - Extentions
-extension UIColor {
-    convenience init(hex: String, alpha: CGFloat = 1.0) {
-        var cString:String = hex.trimmingCharacters(in: .whitespacesAndNewlines).uppercased()
-        
-        if (cString.hasPrefix("#")) {
-            cString.remove(at: cString.startIndex)
-        }
-        
-        if ((cString.count) != 6) {
-            self.init(red: 0, green: 0, blue: 0, alpha: alpha)
-            return
-        }
-        
-        var rgbValue:UInt64 = 0
-        Scanner(string: cString).scanHexInt64(&rgbValue)
-        
-        self.init(
-            red: CGFloat((rgbValue & 0xFF0000) >> 16) / 255.0,
-            green: CGFloat((rgbValue & 0x00FF00) >> 8) / 255.0,
-            blue: CGFloat(rgbValue & 0x0000FF) / 255.0,
-            alpha: alpha
-        )
-    }
-}
